// admin-frontend/src/api/generated/model/userPermissionVO.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Admin Backend API
 * 后台管理系统后端 API 文档，前端可通过 OpenAPI Schema 自动生成接口调用代码
 * OpenAPI spec version: 1.0.0
 */
import type { PermissionOverrideVO } from './permissionOverrideVO';
import type { RoleBaseVO } from './roleBaseVO';
import type { UserGroupPermissionVO } from './userGroupPermissionVO';
import type { UserPermissionItemVO } from './userPermissionItemVO';

/**
 * 用户权限详情
 */
export interface UserPermissionVO {
  /** 用户ID */
  userId?: number;
  /** 用户名 */
  username?: string;
  /** 是否超级管理员 */
  isSuperuser?: boolean;
  /** 用户角色列表 */
  roles?: RoleBaseVO[];
  /** 有效权限列表 */
  effectivePermissions?: UserPermissionItemVO[];
  /** 分组权限列表 */
  groupedPermissions?: UserGroupPermissionVO[];
  /** 权限覆盖列表 */
  overrides?: PermissionOverrideVO[];
}