// admin-frontend/src/api/generated/model/userAvailablePermissionVO.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Admin Backend API
 * 后台管理系统后端 API 文档，前端可通过 OpenAPI Schema 自动生成接口调用代码
 * OpenAPI spec version: 1.0.0
 */
import type { CanOverrideVO } from './canOverrideVO';
import type { UnassignedGroupVO } from './unassignedGroupVO';

/**
 * 用户可用权限（未拥有的）
 */
export interface UserAvailablePermissionVO {
  /** 用户ID */
  userId?: number;
  /** 用户名 */
  username?: string;
  /** 是否超级管理员 */
  isSuperuser?: boolean;
  /** 未分配权限（按分组聚合） */
  unassignedPermissions?: UnassignedGroupVO[];
  /** 可被覆盖的权限（当前被拒绝的） */
  canBeOverridden?: CanOverrideVO[];
}